DRY up Generic model
