+++ app/models/document.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal:true
 
 # Sets the expected behaviors and validations for a document
-class Document < Generic
+class Document < ActiveFedora::Base
   include ::Hyrax::WorkBehavior
-  self.indexer = DocumentIndexer
+  include ::OregonDigital::WorkBehavior
   # Change this to restrict which works can be added as a child.
   # self.valid_child_concerns = []
   validates :title, presence: { message: 'Your work must have a title.' }
