+++ app/models/generic.rb
@@ -3,10 +3,7 @@
 # Sets the expected behaviors of a generic work
 class Generic < ActiveFedora::Base
   include ::Hyrax::WorkBehavior
-  attr_writer :graph_fetch_failures
-
-  after_save :enqueue_fetch_failures
-  before_save :resolve_oembed_errors
+  include ::OregonDigital::WorkBehavior
 
   self.indexer = GenericIndexer
   # Change this to restrict which works can be added as a child.
@@ -16,35 +13,4 @@ class Generic < ActiveFedora::Base
   # This must be included at the end, because it finalizes the metadata
   # schema (by adding accepts_nested_attributes)
   include ::OregonDigital::GenericMetadata
-
-  def graph_fetch_failures
-    @graph_fetch_failures ||= []
-  end
-
-  private
-
-  def enqueue_fetch_failures
-    graph_fetch_failures.uniq.each do |rdf_subject|
-      enqueue_fetch_failure(rdf_subject)
-    end
-  end
-
-  ##
-  # Returns an RDF::Graph that is stored as a placeholder
-  #
-  # @param uri [RDF::Uri] the URI to fetch
-  # @param [User] the user to alert about this failed fetch
-  def enqueue_fetch_failure(uri)
-    user = User.find_by_user_key(depositor)
-    # Email user about failure
-    Hyrax.config.callback.run(:ld_fetch_error, user, uri)
-
-    FetchGraphWorker.perform_in(15.minutes, uri, user)
-  end
-
-  # If the oembed_url changed all previous errors are invalid
-  def resolve_oembed_errors
-    errors = OembedError.find_by(document_id: id)
-    errors.delete if oembed_url_changed? && !errors.blank?
-  end
 end
