+++ app/models/image.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal:true
 
 # Sets the expected behaviors for an image work
-class Image < Generic
+class Image < ActiveFedora::Base
   include ::Hyrax::WorkBehavior
+  include ::OregonDigital::WorkBehavior
 
   self.indexer = ImageIndexer
   # Change this to restrict which works can be added as a child.
