+++ app/models/video.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal:true
 
 # Sets basic behaviors for a video work
-class Video < Generic
+class Video < ActiveFedora::Base
   include ::Hyrax::WorkBehavior
+  include ::OregonDigital::WorkBehavior
 
   self.indexer = VideoIndexer
   # Change this to restrict which works can be added as a child.
